今日舔狗、赛马、聊天机器人、提肛小助手

+++ src/main/java/com/yangye/wechatrobot/plugin/horserace/Horse.java
@@ -26,5 +26,5 @@ public class Horse {
     /**
      * 距离
      */
-    private double distance;
+    private int distance;
 }
